chore: Reverting to 3 compare methods
After discussion with mentors on the knowledge board, I reverted to an old version, so that I was using 3 distinct compare methods, per the criteria. Besides that, no real changes. As I am already over the time and it is working, I am simply going to submit it after I go over all the requirements one more time.